Added Quiz logic and remote model in Clock

+++ Assets/Code/Scripts/Reward System/DTO/RewardScriptableObject.cs
@@ -4,7 +4,7 @@ namespace Project.RewardSystem{
     [CreateAssetMenu(fileName = "RewardData", menuName ="STO/Reward/RewardData")]
     public class RewardScriptableObject : ScriptableObject{
         public AudioClip rewardSound;
-        public float Goal;
+        public int Goal;
         public RewardBadgeSTO Badge;
         public Sprite UnacquiredBadge;
     }
